+++ src/jllvm/vm/CMakeLists.txt
@@ -14,7 +14,7 @@
 llvm_map_components_to_libnames(llvm_native_libs ${LLVM_NATIVE_ARCH})
 
 add_library(JLLVMVirtualMachine VirtualMachine.cpp JIT.cpp StackMapRegistrationPlugin.cpp
-        JNIImplementation.cpp NativeImplementation.cpp StringInterner.cpp MarkSanitizersGCLeafs.cpp native/IO.cpp
+        JNIImplementation.cpp NativeImplementation.cpp MarkSanitizersGCLeafs.cpp native/IO.cpp
         native/Lang.cpp native/JDK.cpp native/Security.cpp
         ClassObjectStubImportPass.cpp)
 target_link_libraries(JLLVMVirtualMachine
